Register views of the `extras` app
