+++ templates/extras/tag/_base.html
@@ -20,7 +20,7 @@
 {% block content %}
 <ul class="nav nav-tabs mb-3">
   <li class="nav-item">
-    <a class="nav-link{% if tab == 'main' %} active{% endif %}" href="{% url 'extras:tag_view' pk=instance.pk %}">
+    <a class="nav-link{% if tab == 'main' %} active{% endif %}" href="{% url 'extras:tag' pk=instance.pk %}">
       <i class="fa-fw fa-solid fa-info-circle"></i> Details
     </a>
   </li>
